requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-talisman==1.1.0
gunicorn==21.2.0
ultralytics==8.0.230
numpy==1.26.3
opencv-python==4.9.0.80
Pillow==10.2.0
google-auth==2.27.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
werkzeug==3.0.1
PyJWT==2.8.0
requests==2.31.0
flask-limiter==3.5.0